pyproject.toml
[project]
name = "secure-net-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
